Recheck same position with new rules
When switching condition after a match,
recheck the same position against the
new condition before to make next move.

Also further clean up a bit previous
commit while there.

+++ src/search.h
@@ -56,7 +56,7 @@ struct Condition {
 
 struct MatchingGame {
   uint64_t gameOfs;
-  std::vector<uint16_t> plies;
+  std::vector<size_t> plies;
 };
 
 struct Data {
